Fix critical training bugs: SFT timestep arg, GRPO response mask, RMSNorm bf16 overflow
- SFT loop: model(input_ids, t) — DIMBA.forward requires timestep tensor (crash fix)
- GRPO _build_response_mask: vectorised per-row length to avoid masking padding zeros
  as response tokens when adaptive_stop generates variable-length completions
- RMSNorm.forward: cast to fp32 before squaring to prevent bf16 overflow (x²>256→inf)
- TokenEmbedding.resize(): add method; sync output_head weight-tying buffer after resize
- GlaiveReasoningDataset: stream 22M-row dataset instead of materialising all rows first
- FrontierCoTMix: per-source try/except so one failed HF download doesn't abort all
- _run_grpo_pass: dataclasses.replace() for safe config copy; fix 'general' label trainer
- SFT LR schedule: use optimizer steps not micro-batch steps for cosine decay
- SFT: flush remaining accumulated gradients at end of each epoch
- MGPO + Long2Short (VibeThinker): new grpo.py with normalized brevity bonus,
  max-entropy prompt weighting, LR warmup, per-token loss normalization
- Two-stage SFT: stage 2 hard subset (Orca-Math + Frontier CoT only)
- Sequential GRPO: math RL → code RL via --grpo-data seq
- HuggingFace auto-upload: --hf-token / --hf-repo uploads after each phase

+++ src/dimba/models/denoiser.py
@@ -37,7 +37,8 @@ class RMSNorm(nn.Module):
         self.eps = eps
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
-        rms = x.pow(2).mean(-1, keepdim=True).add(self.eps).sqrt()
+        # Cast to fp32 for variance: bf16 max is ~65504, so x²>256 overflows to inf.
+        rms = x.float().pow(2).mean(-1, keepdim=True).add(self.eps).sqrt().to(x.dtype)
         return (x / rms) * self.weight
 
 # Resolve the best available Mamba implementation once, at import time.
